[exp, wip] add snapshot of picture loading experiments
The final goal here is to show a picture of the original QR code (from a
photo) and let the user drag the QR code "frame" over the picture in the
correct place.
Later, the user-specified picture area could be displayed in the main
canvas (below the pixel data), as reference for the user.

These are two (alternative) experimental approaches in this snapshot:
- a Canvas area. It shows a QR code frame, and it can show the loaded
  picture.
- an SVG area. It show the QR code frame and some draggable corner markers.

The SVG approach has the advantage that mouse interaction appears to be
much easier than with a Canvas.