implement simple method to allow scrolling outside of picture area
This allows to place the corner markers outside of the picture, in case
parts of the QR code are not within the picture.

The available area to scroll in is simply three times the size of the
picture. This might not be sufficient for all pictures, but it kept the
implementation simple.